refactor(mentoring): URL 수정
