+++ src/app/mentor/layout.tsx
@@ -1,30 +1,14 @@
-'use client'
-
-import { usePathname } from 'next/navigation'
-import {Header} from "@/components/layout/Header";
-import {Sidebar} from "@/components/layout/Sidebar";
 import {Footer} from "@/components/layout/Footer";
+import {BaseHeader} from "@/components/layout/BaseHeader";
 
-export default function MainLayout({ children, }: { children: React.ReactNode }) {
-    const pathname = usePathname()
-    const showSidebar = pathname !== '/'
-
+export default function MentorRegistrationLayout({ children, }: { children: React.ReactNode; }) {
     return (
-        <div className="min-h-screen flex flex-col bg-white">
-            <div className="fixed top-0 left-0 right-0 z-50">
-                <Header/>
-            </div>
-            <div className="flex-1 flex pt-[90px]">
-                {showSidebar && (
-                    <aside className="fixed left-0 w-[80px] h-full -mt-[2px]">
-                        <Sidebar/>
-                    </aside>
-                )}
-                <main className={`flex-1 ${showSidebar ? 'pl-[80px]' : ''}`}>
-                    {children}
-                </main>
-            </div>
-            <Footer/>
+        <div className="flex min-h-screen flex-col bg-white">
+            <BaseHeader />
+            <main className="flex-1 container mx-auto px-4 py-8">
+                {children}
+            </main>
+            <Footer />
         </div>
-    )
+    );
 }
