+++ src/app/mentoring/[id]/page.tsx
